+++ package.json
@@ -12,6 +12,7 @@
 		"prepare": "husky"
 	},
 	"dependencies": {
+		"@aws-amplify/adapter-nextjs": "^1.2.4",
 		"@hookform/resolvers": "^3.6.0",
 		"@radix-ui/react-dialog": "^1.0.5",
 		"@radix-ui/react-label": "^2.0.2",
