+++ src/components/auth/ConfigureAmplifyClientSide.ts
@@ -1,7 +1,7 @@
 'use client';
 
 import { Amplify } from 'aws-amplify';
-import config from '../../aws-exports';
+import config from '@/aws-exports';
 
 Amplify.configure(config, { ssr: true });
