+++ src/utils/amplifyServerUtils.ts
@@ -1,0 +1,6 @@
+import { createServerRunner } from '@aws-amplify/adapter-nextjs';
+import config from '@/aws-exports';
+
+export const { runWithAmplifyServerContext } = createServerRunner({
+	config,
+});
